Include print out for new object in parser.  Fix parser for cooling channel given change of interface.
Former-commit-id: 1b2d8d25d306ae9582ac0eee329eb4f6ccd05644

+++ src/BDSParser.cc
@@ -119,7 +119,7 @@ const GMAD::CoolingChannel* BDSParser::GetCoolingChannel(const std::string& obje
     }
   else
     {
-      if (coolingChannelObjectMap->size() != coolingchannel_list.size())
+      if ((int)coolingChannelObjectMap->size() != coolingchannel_list.size())
 	{
 	  coolingChannelObjectMap->clear();
 	  for (auto& cco: coolingchannel_list)
